pyproject.toml
[project]
name = "llm-research-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.11",
]
